Move time-consuming lerobot tests to integration
