+++ library/tests/unit/policies/test_lerobot.py
@@ -950,537 +950,3 @@ class TestCheckpointConverter:
 
         result = _add_prefix({}, "_lerobot_policy.")
         assert result == {}
-
-
-class TestLeRobotPolicyNumericalEquivalence:
-    """Tests verifying wrapper produces identical results to direct LeRobot calls.
-
-    These tests ensure our wrapper's predict_action_chunk and select_action
-    produce numerically equivalent outputs to calling the underlying LeRobot
-    policy methods directly.
-
-    Uses synthetic mock data to avoid FFmpeg/torchcodec dependency in CI.
-    """
-
-    @pytest.fixture
-    def policy_and_batch(self, lerobot_imports, pusht_dataset):
-        """Create policy and matching synthetic batch on same device.
-
-        Uses ``_make_training_batch`` so the shapes are correct for both
-        inference *and* training (ACT's VAE encoder requires temporal dims).
-        """
-        LeRobotPolicy = lerobot_imports["LeRobotPolicy"]
-
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-        policy = LeRobotPolicy.from_dataset("act", pusht_dataset)
-        policy = policy.to(device)
-        policy.eval()
-
-        batch = _make_training_batch(policy._config, device)
-
-        return policy, batch
-
-    def test_select_action_matches_lerobot_directly(self, policy_and_batch):
-        """Verify wrapper.select_action == lerobot_policy.select_action."""
-        policy, batch = policy_and_batch
-
-        policy.reset()
-
-        preprocessed = policy._preprocessor(_clone_batch(batch))
-        lerobot_action = policy.lerobot_policy.select_action(preprocessed)
-
-        policy.reset()
-
-        wrapper_action = policy.select_action(_clone_batch(batch))
-
-        # Should be numerically identical
-        torch.testing.assert_close(
-            wrapper_action,
-            policy._postprocessor(lerobot_action),
-            rtol=1e-5,
-            atol=1e-5,
-            msg="Wrapper select_action should match LeRobot select_action",
-        )
-
-    def test_predict_action_chunk_matches_lerobot_directly(self, policy_and_batch):
-        """Verify wrapper.predict_action_chunk == lerobot_policy.predict_action_chunk."""
-        policy, batch = policy_and_batch
-
-        policy.reset()
-
-        preprocessed = policy._preprocessor(_clone_batch(batch))
-        lerobot_chunk = policy.lerobot_policy.predict_action_chunk(preprocessed)
-
-        policy.reset()
-
-        wrapper_chunk = policy.predict_action_chunk(_clone_batch(batch))
-
-        # Should be numerically identical
-        torch.testing.assert_close(
-            wrapper_chunk,
-            policy._postprocessor(lerobot_chunk),
-            rtol=1e-5,
-            atol=1e-5,
-            msg="Wrapper predict_action_chunk should match LeRobot predict_action_chunk",
-        )
-
-    def test_forward_training_returns_loss(self, policy_and_batch):
-        """Verify forward in training mode returns (loss, loss_dict)."""
-        policy, batch = policy_and_batch
-
-        policy.train()
-        policy.reset()
-
-        output = policy(_clone_batch(batch))
-
-        assert isinstance(output, tuple)
-        loss = output[0]
-        assert isinstance(loss, torch.Tensor)
-        assert loss.numel() == 1
-
-    def test_select_action_shape_is_single_action(self, policy_and_batch):
-        """Verify select_action returns single action per batch item."""
-        policy, batch = policy_and_batch
-
-        policy.eval()
-        policy.reset()
-        action = policy.select_action(_clone_batch(batch))
-
-        # select_action returns (batch, action_dim) - one action per batch item
-        # LeRobot's select_action preserves the batch dimension
-        action_dim = policy._config.output_features["action"].shape[0]
-        assert action.dim() == 2, f"Expected 2D tensor, got shape {action.shape}"
-        assert action.shape[0] == 1  # batch size
-        assert action.shape[1] == action_dim  # action_dim
-
-    def test_predict_action_chunk_shape_is_full_chunk(self, policy_and_batch):
-        """Verify predict_action_chunk returns full chunk."""
-        policy, batch = policy_and_batch
-
-        policy.eval()
-        policy.reset()
-        chunk = policy.predict_action_chunk(_clone_batch(batch))
-
-        # predict_action_chunk should return (batch, chunk_size, action_dim)
-        action_dim = policy._config.output_features["action"].shape[0]
-        assert chunk.dim() == 3, f"Expected 3D tensor, got shape {chunk.shape}"
-        assert chunk.shape[0] == 1  # batch size
-        assert chunk.shape[1] == policy._config.chunk_size  # chunk_size
-        assert chunk.shape[2] == action_dim  # action_dim
-
-    def test_multiple_select_action_uses_cached_chunk(self, lerobot_imports, pusht_dataset):
-        """Verify select_action uses internal queue correctly."""
-        LeRobotPolicy = lerobot_imports["LeRobotPolicy"]
-
-        device = torch.device("cuda" if torch.cuda.is_available() else "cpu")
-
-        policy = LeRobotPolicy.from_dataset("act", pusht_dataset, n_action_steps=3)
-        policy = policy.to(device)
-        policy.eval()
-
-        batch = _make_training_batch(policy._config, device)
-
-        policy.reset()
-
-        full_chunk = policy.predict_action_chunk(_clone_batch(batch))
-
-        policy.reset()
-        actions = []
-        for _ in range(3):
-            action = policy.select_action(_clone_batch(batch))
-            actions.append(action)
-
-        for action in actions:
-            assert action.shape == (1, full_chunk.shape[2])
-
-
-def _clone_batch(batch: dict[str, Any]) -> dict[str, Any]:
-    """Deep-clone a batch dict, detaching and cloning all tensors."""
-    import copy
-
-    out: dict[str, Any] = {}
-    for key, value in batch.items():
-        if isinstance(value, torch.Tensor):
-            out[key] = value.detach().clone()
-        elif isinstance(value, dict):
-            out[key] = _clone_batch(value)
-        else:
-            out[key] = copy.deepcopy(value)
-    return out
-
-
-def _make_training_batch(config, device: torch.device) -> dict[str, torch.Tensor]:
-    """Build a synthetic training batch with correct shapes for any LeRobot policy.
-
-    Action temporal dimension resolution (in order):
-    - ACT / VLA family: ``chunk_size``
-    - Diffusion: ``horizon``
-    - Fallback: ``(B, action_dim)`` for non-temporal policies.
-
-    Tokenized policies (smolvla, …) detect a tokenizer in the preprocessor
-    pipeline via the presence of language-related config attributes and
-    receive a synthetic ``task`` string.
-    """
-    batch: dict[str, torch.Tensor] = {}
-
-    n_obs_steps = getattr(config, "n_obs_steps", 1)
-    for key, feature in config.input_features.items():
-        if n_obs_steps > 1:
-            batch[key] = torch.randn(1, n_obs_steps, *feature.shape, device=device)
-        else:
-            batch[key] = torch.randn(1, *feature.shape, device=device)
-
-    action_dim = config.output_features["action"].shape[0]
-
-    chunk = getattr(config, "chunk_size", None)
-    horizon = getattr(config, "horizon", None)
-    temporal_len = chunk or horizon
-
-    if temporal_len is not None:
-        batch["action"] = torch.randn(1, temporal_len, action_dim, device=device)
-        batch["action_is_pad"] = torch.zeros(1, temporal_len, dtype=torch.bool, device=device)
-    else:
-        batch["action"] = torch.randn(1, action_dim, device=device)
-        batch["action_is_pad"] = torch.zeros(1, dtype=torch.bool, device=device)
-
-    if any(hasattr(config, attr) for attr in ("tokenizer_max_length", "max_state_dim", "vlm_model_name")):
-        batch["task"] = ["pick up the block"]
-
-    return batch
-
-
-_TRAINING_POLICY_NAMES = ["act", "diffusion", "smolvla"]
-"""Policies validated end-to-end on CPU with the pusht dataset fixture.
-
-Excluded with reason:
-- ``vlas`` (pi0/pi05/pi0_fast/groot): require GPU + bf16; integration tier only.
-- ``xvla``: requires explicit ``vision_config`` kwarg, not derivable from the dataset.
-"""
-
-
-class TestTrainingNumericalEquivalence:
-    """Verify wrapper training produces identical loss to calling inner LeRobot policy directly.
-
-    The wrapper's forward() applies _preprocessor then delegates to lerobot_policy().
-    These tests confirm no silent modifications occur during that delegation, so
-    training through the wrapper is numerically equivalent to native LeRobot training.
-    """
-
-    @pytest.fixture(params=_TRAINING_POLICY_NAMES)
-    def training_policy_and_batch(self, request, lerobot_imports, pusht_dataset):
-        policy_name = request.param
-        LeRobotPolicy = lerobot_imports["LeRobotPolicy"]
-
-        device = torch.device("cpu")
-
-        # Pass device="cpu" so the LeRobot config (and its DeviceProcessorStep)
-        # targets CPU; otherwise LeRobot autodetects CUDA when available and the
-        # preprocessor moves the batch to GPU while wrapper weights stay on CPU.
-        policy = LeRobotPolicy.from_dataset(policy_name, pusht_dataset, device="cpu")
-        policy.train()
-
-        batch = _make_training_batch(policy._config, device)
-
-        return policy, batch, policy_name
-
-    def test_wrapper_forward_matches_native_path(self, training_policy_and_batch):
-        """Wrapper.forward(batch) must produce the same loss as preprocessor + native call.
-
-        This is the core wrapper-vs-native equivalence guarantee at the unit level:
-        running the batch through ``policy(batch)`` (which preprocesses internally)
-        must be numerically identical to manually preprocessing and calling the
-        underlying ``lerobot_policy`` directly with the same seed.
-        """
-        policy, batch, _ = training_policy_and_batch
-
-        preprocessed = policy._preprocessor(_clone_batch(batch))
-
-        torch.manual_seed(42)
-        native_loss, _ = policy.lerobot_policy(preprocessed)
-
-        torch.manual_seed(42)
-        wrapper_loss, _ = policy(_clone_batch(batch))
-
-        torch.testing.assert_close(wrapper_loss, native_loss, rtol=1e-6, atol=1e-6)
-
-    def test_wrapper_forward_produces_loss_and_dict(self, training_policy_and_batch):
-        policy, batch, _ = training_policy_and_batch
-
-        output = policy(batch)
-
-        assert isinstance(output, tuple)
-        assert len(output) == 2
-        loss, _ = output
-        assert isinstance(loss, torch.Tensor)
-        assert loss.ndim == 0
-        assert loss.requires_grad
-
-    def test_act_loss_dict_contains_expected_keys(self, training_policy_and_batch):
-        policy, batch, policy_name = training_policy_and_batch
-        if policy_name != "act":
-            pytest.skip("ACT-specific: loss_dict structure")
-
-        _, loss_dict = policy(batch)
-
-        assert loss_dict is not None
-        assert "l1_loss" in loss_dict
-
-    def test_diffusion_loss_dict_is_none(self, training_policy_and_batch):
-        policy, batch, policy_name = training_policy_and_batch
-        if policy_name != "diffusion":
-            pytest.skip("Diffusion-specific: loss_dict is None")
-
-        _, loss_dict = policy(batch)
-
-        assert loss_dict is None
-
-    def test_gradient_flows_through_wrapper(self, training_policy_and_batch):
-        policy, batch, _ = training_policy_and_batch
-
-        loss, _ = policy(batch)
-        loss.backward()
-
-        params_with_grad = [
-            name for name, p in policy.lerobot_policy.named_parameters() if p.requires_grad and p.grad is not None
-        ]
-        total_params = [name for name, p in policy.lerobot_policy.named_parameters() if p.requires_grad]
-
-        assert len(params_with_grad) > 0
-        assert len(params_with_grad) == len(total_params)
-
-    def test_optimizer_step_updates_weights(self, training_policy_and_batch):
-        policy, batch, _ = training_policy_and_batch
-
-        optimizer = policy.configure_optimizers()
-
-        before = {n: p.clone() for n, p in policy.lerobot_policy.named_parameters() if p.requires_grad}
-
-        optimizer.zero_grad()
-        loss, _ = policy(batch)
-        loss.backward()
-        optimizer.step()
-
-        changed = 0
-        for name, p in policy.lerobot_policy.named_parameters():
-            if p.requires_grad and not torch.equal(p, before[name]):
-                changed += 1
-
-        assert changed > 0
-
-    def test_wrapper_and_direct_gradient_match(self, training_policy_and_batch):
-        policy, batch, _ = training_policy_and_batch
-
-        # Clone batch before each call to avoid in-place mutation
-        preprocessed_a = policy._preprocessor(_clone_batch(batch))
-
-        # --- Direct path: call inner LeRobot policy on preprocessed data ---
-        policy.zero_grad()
-        torch.manual_seed(99)
-        direct_loss, _ = policy.lerobot_policy(preprocessed_a)
-        direct_loss.backward()
-        direct_grads = {
-            n: p.grad.clone()
-            for n, p in policy.lerobot_policy.named_parameters()
-            if p.requires_grad and p.grad is not None
-        }
-
-        # --- Wrapper path: call policy(batch) which preprocesses internally ---
-        policy.zero_grad()
-        torch.manual_seed(99)
-        wrapper_loss, _ = policy(_clone_batch(batch))
-        wrapper_loss.backward()
-        wrapper_grads = {
-            n: p.grad.clone()
-            for n, p in policy.lerobot_policy.named_parameters()
-            if p.requires_grad and p.grad is not None
-        }
-
-        assert set(direct_grads.keys()) == set(wrapper_grads.keys())
-        for name in direct_grads:
-            torch.testing.assert_close(
-                direct_grads[name],
-                wrapper_grads[name],
-                rtol=1e-6,
-                atol=1e-6,
-            )
-
-    def test_preprocessing_modifies_values(self, training_policy_and_batch):
-        policy, batch, _ = training_policy_and_batch
-
-        if policy._dataset_stats is None:
-            pytest.skip("No dataset_stats; preprocessor cannot normalize")
-
-        raw_batch = _clone_batch(batch)
-        preprocessed = policy._preprocessor(_clone_batch(batch))
-
-        modified_keys = [
-            key for key in policy._config.input_features if not torch.equal(raw_batch[key], preprocessed[key])
-        ]
-        assert modified_keys, (
-            f"Preprocessing did not modify any input feature; expected at least one of "
-            f"{list(policy._config.input_features)} to be normalized."
-        )
-
-    def test_loss_finite_and_positive(self, training_policy_and_batch):
-        policy, batch, _ = training_policy_and_batch
-
-        loss, _ = policy(batch)
-
-        assert torch.isfinite(loss)
-        assert loss.item() >= 0
-
-
-class TestMultiStepTrainingTrajectory:
-    """Verify multi-step training through the wrapper matches native LeRobot exactly.
-
-    Runs N training steps with identical weights, optimizer, and batch through both
-    the physicalai wrapper and bare LeRobot policy, then asserts:
-    - Per-step loss values are identical (atol=0)
-    - Loss decreases from first to last step
-    - Final model weights are identical
-    """
-
-    NUM_STEPS = 5
-
-    @pytest.fixture(params=_TRAINING_POLICY_NAMES)
-    def dual_setup(self, request, lerobot_imports, pusht_dataset):
-        """Create wrapper + native LeRobot policy with identical weights and optimizers."""
-        import copy
-
-        policy_name = request.param
-        LeRobotPolicy = lerobot_imports["LeRobotPolicy"]
-
-        device = torch.device("cpu")
-
-        # See TestTrainingNumericalEquivalence.training_policy_and_batch for why
-        # device="cpu" must be passed as a config kwarg, not via .to("cpu").
-        wrapper = LeRobotPolicy.from_dataset(policy_name, pusht_dataset, device="cpu")
-        wrapper.train()
-
-        # --- Native path: deep-copy the inner LeRobot policy so weights are identical ---
-        native = copy.deepcopy(wrapper.lerobot_policy)
-        native = native.to(device)
-        native.train()
-
-        # Build identical optimizers using LeRobot's own preset
-        wrapper_optimizer = wrapper.configure_optimizers()
-        native_optimizer = wrapper._config.get_optimizer_preset().build(native.get_optim_params())
-
-        # Build a fixed training batch
-        batch = _make_training_batch(wrapper._config, device)
-
-        return {
-            "wrapper": wrapper,
-            "native": native,
-            "wrapper_optimizer": wrapper_optimizer,
-            "native_optimizer": native_optimizer,
-            "batch": batch,
-            "policy_name": policy_name,
-        }
-
-    def test_loss_trajectories_match(self, dual_setup):
-        """Per-step loss values from wrapper and native must be bit-identical."""
-        wrapper = dual_setup["wrapper"]
-        native = dual_setup["native"]
-        w_opt = dual_setup["wrapper_optimizer"]
-        n_opt = dual_setup["native_optimizer"]
-        batch = dual_setup["batch"]
-
-        wrapper_losses: list[float] = []
-        native_losses: list[float] = []
-
-        for step in range(self.NUM_STEPS):
-            w_opt.zero_grad()
-            torch.manual_seed(step)
-            w_loss, _ = wrapper(_clone_batch(batch))
-            w_loss.backward()
-            w_opt.step()
-            wrapper_losses.append(w_loss.item())
-
-            n_opt.zero_grad()
-            preprocessed = wrapper._preprocessor(_clone_batch(batch))
-            torch.manual_seed(step)
-            n_loss, _ = native(preprocessed)
-            n_loss.backward()
-            n_opt.step()
-            native_losses.append(n_loss.item())
-
-        for step, (wl, nl) in enumerate(zip(wrapper_losses, native_losses)):
-            torch.testing.assert_close(
-                torch.tensor(wl),
-                torch.tensor(nl),
-                rtol=1e-6,
-                atol=1e-6,
-                msg=lambda m, step=step, wl=wl, nl=nl: f"Step {step}: wrapper loss {wl} != native loss {nl} ({m})",
-            )
-
-    def test_loss_decreases_over_training(self, dual_setup):
-        """Average loss in later steps should be lower than early steps.
-
-        Diffusion policies have inherently noisy loss (random noise + timesteps
-        each forward), so we compare the mean of the first half vs the second
-        half over more steps rather than demanding strict monotonic decrease.
-        """
-        wrapper = dual_setup["wrapper"]
-        w_opt = dual_setup["wrapper_optimizer"]
-        batch = dual_setup["batch"]
-
-        num_steps = 20
-        losses: list[float] = []
-        for step in range(num_steps):
-            w_opt.zero_grad()
-            torch.manual_seed(step)
-            loss, _ = wrapper(_clone_batch(batch))
-            loss.backward()
-            w_opt.step()
-            losses.append(loss.item())
-
-        mid = num_steps // 2
-        first_half_mean = sum(losses[:mid]) / mid
-        second_half_mean = sum(losses[mid:]) / (num_steps - mid)
-
-        assert first_half_mean > second_half_mean, (
-            f"Loss did not decrease: first_half_mean={first_half_mean:.6f}, "
-            f"second_half_mean={second_half_mean:.6f}, "
-            f"trajectory={[f'{val:.4f}' for val in losses]}"
-        )
-
-    def test_final_weights_match(self, dual_setup):
-        """After N identical steps, wrapper and native should have identical parameters."""
-        wrapper = dual_setup["wrapper"]
-        native = dual_setup["native"]
-        w_opt = dual_setup["wrapper_optimizer"]
-        n_opt = dual_setup["native_optimizer"]
-        batch = dual_setup["batch"]
-
-        for step in range(self.NUM_STEPS):
-            w_opt.zero_grad()
-            torch.manual_seed(step)
-            w_loss, _ = wrapper(_clone_batch(batch))
-            w_loss.backward()
-            w_opt.step()
-
-            n_opt.zero_grad()
-            preprocessed = wrapper._preprocessor(_clone_batch(batch))
-            torch.manual_seed(step)
-            n_loss, _ = native(preprocessed)
-            n_loss.backward()
-            n_opt.step()
-
-        # Compare all parameters
-        wrapper_params = dict(wrapper.lerobot_policy.named_parameters())
-        native_params = dict(native.named_parameters())
-
-        assert set(wrapper_params.keys()) == set(native_params.keys()), (
-            "Parameter names differ between wrapper and native"
-        )
-
-        for name in wrapper_params:
-            torch.testing.assert_close(
-                wrapper_params[name],
-                native_params[name],
-                rtol=1e-6,
-                atol=1e-6,
-                msg=f"Parameter '{name}' diverged after {self.NUM_STEPS} training steps",
-            )
